zimkag_email_watcher/requirements.txt
# Gmail API
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# HTTP client → ZimKAG webapp
requests==2.32.3

# Config
python-dotenv==1.0.1

# Document parsing (pre-filter)
pdfplumber==0.11.4
python-docx==1.1.2
